Rewrite display module
The html elements are now editable using the contenteditable
attribute. This resulted in much cleaner code. Additional features
added from the last commit are error messages shown when invalid data
is sent by a user editing a todo, and a discard changes option when a
user is editing a todo.