Refactor todoDisplay and application
Added a data object to the HTML element objects to store additional
information about their current state. This meant the code now does
not rely as much on closures and meant that the factories creating the
display elements could be broken down into smaller pieces. Modified
the application module to be compatiable with these new changes.